Bug fix: crashes when reading damaged images, caused by buffer overruns, or by deleting uninitialized pointers (found by Apple).

+++ OpenEXR/IlmImf/ImfPxr24Compressor.h
@@ -51,7 +51,10 @@ class Pxr24Compressor: public Compressor
 {
   public:
 
-    Pxr24Compressor (const Header &hdr, int maxScanLineSize, int numScanLines);
+    Pxr24Compressor (const Header &hdr, 
+                     size_t maxScanLineSize,
+                     size_t numScanLines);
+
     virtual ~Pxr24Compressor ();
 
     virtual int		numScanLines () const;
